backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
pydantic>=2.9.0
sqlalchemy>=2.0.23
mysql-connector-python>=9.0.0
python-dotenv>=1.0.0
